requirements.txt
astroid==2.3.3
beautifulsoup4==4.9.0
bs4==0.0.1
DateTime==4.3
IMDbPY==6.8
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
numpy==1.18.2
ordered-set==3.1.1
pandas==1.0.3
pkg-resources==0.0.0
PyLaTeX==1.3.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.15
textable==0.5.2
texttable==1.6.2
typed-ast==1.4.1
wrapt==1.11.2
zope.interface==5.0.2
